@@ -1,6 +1,6 @@
 module github.com/fairwindsops/pluto/v5
 
-go 1.25.5
+go 1.26.0
 
 require (
 	github.com/olekukonko/tablewriter v1.1.3
