@@ -16,6 +16,7 @@ package api
 
 import (
 	"os"
+	"slices"
 )
 
 // IsFileOrStdin detects if a file exists, or returns true if - is passed
@@ -32,10 +33,5 @@ func IsFileOrStdin(name string) bool {
 
 // StringInSlice returns true if the string is contained in the slice
 func StringInSlice(s string, slice []string) bool {
-	for _, v := range slice {
-		if s == v {
-			return true
-		}
-	}
-	return false
+	return slices.Contains(slice, s)
 }
